Added matmult_p
Need to make Multiply.c, created the shared memory object and decided to pass the address to the child process. That way the child is able to write to the shared memory. This may not be the best way of doing it, I believe that the child will have the ability to access the shared memory regardless. Need to read more on shmget().